[feature] Add score, viewcounts and comments counts displaying
