+++ app/src/main/java/com/makentoshe/habrachan/AppActivity.kt
@@ -25,9 +25,9 @@ class AppActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         if (savedInstanceState == null) {
-//            val screen = MainFlowScreen()
+            val screen = MainFlowScreen()
 //            val screen = CommentsScreen(485990)
-            val screen = PostScreen(397271)
+//            val screen = PostScreen(397271)
 //            val screen = PostImageScreen("https://habrastorage.org/webt/r7/i1/o6/r7i1o6qrcdmwgj0nrtzc8ctltvs.jpeg")
             router.newRootScreen(screen)
         }
